+++ FEM/examples/src/heatVec.cpp
@@ -87,7 +87,7 @@ bool HeatVec<dim>::preComputeVec(const VECType* in,VECType* out, double scale)
 {
 
     // apply boundary conditions.
-    std::vector<unsigned int> bdyIndex;
+    std::vector<size_t> bdyIndex;
     m_uiOctDA->getBoundaryNodeIndices(bdyIndex);
 
     for(unsigned int i=0;i<bdyIndex.size();i++)
@@ -98,7 +98,7 @@ template <unsigned int dim>
 bool HeatVec<dim>::postComputeVec(const VECType* in,VECType* out, double scale) {
 
     // apply boundary conditions.
-    std::vector<unsigned int> bdyIndex;
+    std::vector<size_t> bdyIndex;
     m_uiOctDA->getBoundaryNodeIndices(bdyIndex);
 
     for(unsigned int i=0;i<bdyIndex.size();i++)
